Started interface implementation

+++ FourZug/FourZug/Backend/TreeManager/ITreeManager.cs
@@ -5,6 +5,6 @@
     public interface ITreeManager
     {
         // Starts the Bot and returns best move results
-        byte StartBot(string[,] grid, string currentTurn);
+        int StartBot(string[,] grid, string currentTurn);
     }
 }
